Remove `ArrayExt` (move the methods to `Array`)
